refactor: extract function to insert dom elements
